refactor(omegaform): extract persistency composable
